[2.0] Entity fixes (#1231)
* Entity and BlockEntity Serializer fixes

* Chest and Bed fixes

* Remove Log annotation and fix package name inclusion

* bump protocol lib version

* Barrel updates

* Beacon fixes

* Furnace fixes

* revert saveClientData change since items do not need to be sent in NBT

* Enchantment Table fixes

* Add Convenience methods for `setBlockDataAt` which accepts `Vector3i` position

* Furnace Fixes

* Fix build from Porky's lib change

* fix implements clause

+++ src/main/java/cn/nukkit/inventory/ContainerInventory.java
@@ -37,7 +37,7 @@ public abstract class ContainerInventory extends BaseInventory {
     public void onOpen(Player who) {
         super.onOpen(who);
         ContainerOpenPacket packet = new ContainerOpenPacket();
-        packet.setWindowId((byte) who.getWindowId(this));
+        packet.setWindowId(who.getWindowId(this));
         packet.setType((byte) this.getType().getNetworkType());
         InventoryHolder holder = this.getHolder();
         if (holder instanceof BlockEntity) {
@@ -54,7 +54,7 @@ public abstract class ContainerInventory extends BaseInventory {
     @Override
     public void onClose(Player who) {
         ContainerClosePacket packet = new ContainerClosePacket();
-        packet.setWindowId((byte) who.getWindowId(this));
+        packet.setWindowId(who.getWindowId(this));
         who.sendPacket(packet);
         super.onClose(who);
     }
